+++ src/schemathesis/specs/graphql/loaders.py
@@ -35,7 +35,7 @@ def from_path(
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
     encoding: str = "utf8",
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema via a file from an OS path.
 
@@ -51,7 +51,7 @@ def from_path(
             code_sample_style=code_sample_style,
             location=pathlib.Path(path).absolute().as_uri(),
             rate_limit=rate_limit,
-            mask_sensitive_output=mask_sensitive_output,
+            sanitize_output=sanitize_output,
         )
 
 
@@ -79,7 +79,7 @@ def from_url(
     code_sample_style: str = CodeSampleStyle.default().name,
     wait_for_schema: Optional[float] = None,
     rate_limit: Optional[str] = None,
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from the network.
@@ -122,7 +122,7 @@ def from_url(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
-        mask_sensitive_output=mask_sensitive_output,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -135,7 +135,7 @@ def from_file(
     code_sample_style: str = CodeSampleStyle.default().name,
     location: Optional[str] = None,
     rate_limit: Optional[str] = None,
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a file descriptor or a string.
 
@@ -162,7 +162,7 @@ def from_file(
         code_sample_style=code_sample_style,
         location=location,
         rate_limit=rate_limit,
-        mask_sensitive_output=mask_sensitive_output,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -175,7 +175,7 @@ def from_dict(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a Python dictionary.
 
@@ -201,7 +201,7 @@ def from_dict(
         data_generation_methods=prepare_data_generation_methods(data_generation_methods),
         code_sample_style=_code_sample_style,
         rate_limiter=rate_limiter,
-        mask_sensitive_output=mask_sensitive_output,
+        sanitize_output=sanitize_output,
     )  # type: ignore
     dispatch("after_load_schema", hook_context, instance)
     return instance
@@ -215,7 +215,7 @@ def from_wsgi(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a WSGI app.
@@ -239,7 +239,7 @@ def from_wsgi(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
-        mask_sensitive_output=mask_sensitive_output,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -251,7 +251,7 @@ def from_asgi(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
-    mask_sensitive_output: bool = True,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from an ASGI app.
@@ -274,7 +274,7 @@ def from_asgi(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
-        mask_sensitive_output=mask_sensitive_output,
+        sanitize_output=sanitize_output,
     )